Process/requirements.txt
fastapi=0.95.0
uvicorn[standard]=0.21.0
httpx=0.24.0
cachetools=5.3.0
pydantic-settings=2.0.0 
python-dotenv=1.0.0